Sync: john-ThinkPad-E420

+++ parser3/benchmarkset.sh
@@ -1,8 +1,9 @@
 do_one () {
 	f=output/set_$1_$2_$3_$4.err
 
-	if [ ! -e "$f" ]
-	 then
+	if [ ! \( -e "$f" -a -z "$BENCHMARK_FORCE_RECOMPUTE" \) ]
+	then
+		echo Making $f
 		echo "$1 $2 $3 $4" > output/inprogress
 		ocaml make_random_formulas_$1.ml $2 $3 $4 | /usr/bin/time ./main 2> output/set_$1_$2_$3_$4.err > output/set_$1_$2_$3_$4.out
 		rm output/inprogress
@@ -11,11 +12,12 @@ do_one () {
 mkdir -p output 
 
 sudo cpufreq-set -g performance
-cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_available_governors
+#cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_available_governors
 
 if [ -e output/inprogress ]
 then 
-	do_one `cat output/inprogress`
+	( BENCHMARK_FORCE_RECOMPUTE=y
+	do_one `cat output/inprogress` )
 fi
 
 for type in 6sX 6s
